Added Cone Button [WIP]
This is still a work in progress! I am quitting for now.

What works:
It creates a south facing cone based on Fire1.

What doesn't work:
It can't create a cone on different locations
It can't create a cone in different directions

I will work on this another time, probably friday.
It was quite a struggle to get it this far